test: cover parsing edges and split test upstream fakes
