+++ tests/jsonrpc/test_opencode_session_extension_commands.py
@@ -1,9 +1,6 @@
 import httpx
 import pytest
 
-from tests.support.helpers import (
-    DummySessionQueryOpencodeUpstreamClient as DummyOpencodeUpstreamClient,
-)
 from tests.support.helpers import make_basic_auth_header
 from tests.support.jsonrpc_error_assertions import (
     assert_v1_error_context,
@@ -15,6 +12,9 @@ from tests.support.session_extensions import (
     _extension_headers,
     _jsonrpc_app,
     _session_meta,
+)
+from tests.support.session_query_client import (
+    DummySessionQueryOpencodeUpstreamClient as DummyOpencodeUpstreamClient,
 )
 from tests.support.settings import make_settings
